feat: Add ability to edit and delete projects

+++ components/app/task-edit.tsx
@@ -3,17 +3,16 @@
 import { TaskFormSchema, TaskFormType, Task } from "@/lib/definitions";
 import { zodResolver } from "@hookform/resolvers/zod"
 import { useForm } from "react-hook-form";
-import React, { Dispatch, useContext } from "react";
+import React, { Dispatch, useContext, useState } from "react";
 import { deleteTask, updateTask } from "@/lib/actions";
 import { TasksDispatchContext } from "../providers/TasksContext";
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../ui/dialog";
+import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from "../ui/dialog";
 import TaskForm from "./task-form";
 import { useIsMobile } from "@/hooks/use-mobile";
 import { Drawer, DrawerContent, DrawerHeader, DrawerTitle } from "../ui/drawer";
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "../ui/dropdown-menu";
-import { Ellipsis } from "lucide-react";
 import DeleteButton from "./delete-button";
 import { Button } from "../ui/button";
+import { Trash2 } from "lucide-react";
 
 export default function TaskEdit({
     task,
@@ -53,16 +52,29 @@ export default function TaskEdit({
     }
 
     const isMobile = useIsMobile();
+    const [ deleteConfirmOpen, setDeleteConfirmOpen ] = useState(false);
 
     if (isMobile) return (
         <Drawer open={open} onOpenChange={setOpen}>
             <DrawerContent className="px-4 mb-6 box-border sm:[&>button:last-child]:hidden focus-within:outline-none">
                 <DrawerHeader className="px-0 mb-2 flex-row justify-between items-center">
                     <DrawerTitle>Edit task</DrawerTitle>
-                    <DeleteButton onClick={async () => {
-                        if (dispatch) dispatch({ type: "delete", id: task.id });
-                        await deleteTask(task.id);
-                    }} display="icon" />
+                    <Dialog open={deleteConfirmOpen} onOpenChange={setDeleteConfirmOpen}>
+                        <DialogTrigger asChild>
+                            <Button variant="ghost"><Trash2 /></Button>
+                        </DialogTrigger>
+                        <DialogContent>
+                            <DialogTitle>Are you sure?</DialogTitle>
+                            <DialogFooter className="flex-row items-center justify-end gap-2">
+                                <Button variant="secondary" onClick={() => setDeleteConfirmOpen(false)}>Cancel</Button>
+                                <Button variant="destructive" onClick={async () => {
+                                    if (dispatch) dispatch({ type: "delete", id: task.id });
+                                    await deleteTask(task.id);
+                                    setDeleteConfirmOpen(false);
+                                }}>Delete</Button>
+                            </DialogFooter>
+                        </DialogContent>
+                    </Dialog>
                 </DrawerHeader>
                 <TaskForm form={form} onSubmit={onSubmit} close={() => setOpen(false)} />
             </DrawerContent>
@@ -71,17 +83,11 @@ export default function TaskEdit({
 
     return (
         <Dialog open={open} onOpenChange={setOpen}>
-            {/* "[&>button:last-child]:hidden" hides close button */}
-            <DialogContent onOpenAutoFocus={(e) => e.preventDefault()}
-                className="sm:[&>button:last-child]:hidden focus-within:outline-none">
+            <DialogContent>
                 <DialogHeader className="mb-2 flex-row justify-between items-center">
                     <DialogTitle>
                         Edit task
                     </DialogTitle>
-                    <DeleteButton onClick={async () => {
-                        if (dispatch) dispatch({ type: "delete", id: task.id });
-                        await deleteTask(task.id);
-                    }} display="icon" />
                 </DialogHeader>
                 <TaskForm form={form} onSubmit={onSubmit} close={() => setOpen(false)} />
             </DialogContent>
